Move tunable classes to types, lazy import module
