+++ scvi/autotune/__init__.py
@@ -1,5 +1,16 @@
+import warnings
+
+from scvi import settings
+
+try:
+    from ray import tune
+except ModuleNotFoundError as e:
+    raise ModuleNotFoundError(
+        "Please install ray[tune] to use scvi.autotune. Skipping import."
+    ) from e
+
+
 from ._manager import TuneAnalysis, TunerManager
 from ._tuner import ModelTuner
-from ._types import Tunable, TunableMixin
 
-__all__ = ["ModelTuner", "Tunable", "TunableMixin", "TunerManager", "TuneAnalysis"]
+__all__ = ["ModelTuner", "TunerManager", "TuneAnalysis"]
